scaffolding; qubes os dispo multi vm file supports
spent some time looking for log functions, girara_{log,debug,error}
seems to be where I should pick up. The aim of this series of changes
here is to modify zathura to send a msg to a unix socket including the
name of the file it is opening so It can handle loading in that book
from the remote file storage virtual machine (non-disposable). I tried
using the inotify Linux API and it wasn't working the way I wanted it to
so I decided this is the best course of action.