fns:send_bookname_qubes num_bytes_u8: err handling
made everything indicate wether or not it err'ed.
send_bookname_qubes sends the bookname to the qubes-zathura-bookmark
client side to request the book from the server side. num_bytes_u8
takes a null-terminated unsigned char * and returns it's length.
I didn't use strlen because I am using u8's on rust side of things;
I believe that char is signed on x86_64 architectures which is the
intended target.